requirements.txt
tensorflow
numpy
pillow
opencv-python
streamlit
scikit-learn
